Refactor Autonomous modes to use Base Class
Refactored existing OpModes

Added new OpMode called Auto Turn Around
